moving IndexEntry and calling blob #429
