+++ medicines/doc-index-updater/src/create_manager/search_index.rs
@@ -1,12 +1,15 @@
-use crate::{create_manager::models::BlobMetadata, models::IndexEntry, search_client};
+use crate::{
+    create_manager::models::{BlobMetadata, IndexEntry},
+    search_client,
+};
 
 //use std::time::{Duration, Instant};
 
 pub async fn add_to_search_index(
     search_client: &search_client::AzureSearchClient,
     blob_name: &str,
-    _blob: BlobMetadata,
-    _file_size: usize,
+    blob: BlobMetadata,
+    file_size: usize,
 ) -> Result<(), anyhow::Error> {
     let _storage_account =
         std::env::var("STORAGE_ACCOUNT").expect("Set env variable STORAGE_ACCOUNT first!");
@@ -17,7 +20,7 @@ pub async fn add_to_search_index(
     //     &storage_account, &container_name, &blob_name
     // );
 
-    let entry = IndexEntry::for_blob(blob_name.to_owned());
+    let entry = IndexEntry::for_blob(blob_name.to_owned(), blob, file_size);
 
     tracing::info!("Creating index entry ({:?})", entry);
     search_client.create(entry).await.map_err(|e| {
